Move example package functions to separate submodule for better structure
