+++ scopesim/server/__init__.py
@@ -1,4 +1,6 @@
 from .database import (download_packages,
                        list_packages,
                        download_example_data,
-                       list_example_data)
+                       list_example_data,
+                       get_all_packages_on_server)
+from .example_data_utils import download_example_data, list_example_data
